+++ src/features/charts/components/rangeButtons.tsx
@@ -1,6 +1,6 @@
 import React, { memo, FC } from 'react'
 import styled from '@emotion/styled'
-import { Range } from '../models'
+import { Range } from '../models/range'
 
 const LabelRange = styled.div`
     margin: 0rem 0rem 1rem 0.5rem;
@@ -47,12 +47,12 @@ const RangeButton: FC<RangeButtonProps> = ({ range, update, current }) => {
 
 const ranges: Range[] = ['MAX', '5y', '1y', '1m', '5d', '1d'];
 
-type RangeProps = {
+type TestProps = {
     range: Range,
     update: (range: Range) => void
 }
 
-export const RangeButtons = memo<RangeProps>(({ range, update }) => {
+export const RangeButtons = memo<TestProps>(({ range, update }) => {
 
     const buttons = ranges.map(rangeItem => <RangeButton key={rangeItem} current={rangeItem === range} range={rangeItem} update={update} />);
