+++ src/features/companyOverview/index.tsx
@@ -1,10 +1,10 @@
 import React, { FC } from 'react';
 import { useSelector } from 'react-redux';
-import styled from '@emotion/styled'
 import { CompanyOverview } from './models'
-import { AppState } from '../../models';
+import { AppState } from '../../models/appState';
+import styled from '@emotion/styled'
 import { Title } from '../../Root'
-import { Loader } from '../loader'
+import { Loader } from '../loader/Loader'
 
 const CompanyOverviewContainer = styled.div`
     max-height: 400px;
