+++ src/features/keystats/redux/actions.ts
@@ -1,5 +1,5 @@
 import { createAction } from 'redux-actions'
-import { KeyStats } from '../models'
+import { KeyStats } from '../models/keyStats'
 
 export const UPDATE_KEY_STATS = 'UPDATE_KEY_STATS';
 export const updateKeyStats = createAction<KeyStats>(UPDATE_KEY_STATS)
