+++ src/features/news/index.tsx
@@ -1,14 +1,15 @@
 import React, { FC } from 'react';
 import { useSelector } from 'react-redux';
-import styled from '@emotion/styled'
 import { Article, News } from './models'
-import { AppState, Error } from '../../models';
+import { AppState } from '../../models/appState';
+import { Error } from '../../models/errors';
+import styled from '@emotion/styled'
 import { Title } from '../../Root'
 import { ArticleComponent } from './article';
-import { Loader } from '../loader'
+import { Loader } from '../loader/Loader'
 
 
-const NewsLayoutContainer = styled.div`
+const NewsLayoutContainer =styled.div`
     flex: 0 1 34%;
     margin-top: 40px;
     margin-left: 26px;
@@ -34,7 +35,7 @@ const ArticleLayoutContainer = styled.div`
     overflow: auto;
 `
 
-export const NewsComponent: FC = () => {
+export const NewsComponent: FC<{}> = () => {
 
     const news: News = useSelector(({ news }: AppState) => news)
