+++ src/features/search/redux/reducer.ts
@@ -1,5 +1,5 @@
-import { Reducer } from 'redux';
 import { UpdateTickerAction, UPDATE_TICKER } from './actions';
+import { Reducer } from 'redux';
 
 export const search: Reducer<string, UpdateTickerAction> = (
     state = 'aapl',
