--- src/models/index.ts
@@ -1,3 +1,0 @@
-export * from './errors'
-export * from './prices'
-export * from './appState'
