--- src/redux/reducers/index.ts
@@ -1,2 +1,0 @@
-export * from './price'
-export * from './errors'
