+++ tsconfig.json
@@ -17,11 +17,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react",
-    "baseUrl": "./src",
-    "paths": {
-      "features/*": ["features/*"]
-    }
+    "jsx": "react"
   },
   "include": [
     "src"
